backend/requirements.txt
fastapi
uvicorn
requests
pydantic
pillow
numpy
opencv-python
dlib
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
watchdog
google-auth
google-auth-oauthlib
google-api-python-client
watchdog
qrcode
